Phase 3: Stats Protocol + SQLite/DynamoDB adapter split
Split the monolithic stats.py (1453 lines) into a focused stats/
package with a real seam between adapters and consumers:

  src/rushti/stats/
    __init__.py    - re-exports for backwards compat
    paths.py       - DEFAULT_* constants, get_db_path, get_stats_backend
    signature.py   - calculate_task_signature
    sqlite.py      - StatsDatabase (SQLite adapter)
    dynamodb.py    - DynamoDBStatsDatabase (DynamoDB adapter)
    repository.py  - StatsRepository Protocol + create_stats_database

StatsRepository is a runtime_checkable typing.Protocol covering the
write path (start_run, record_task, batch_record_tasks, complete_run,
cleanup_old_data), the read path (get_task_history, get_run_*, etc.),
and the lifecycle (close, __enter__, __exit__). Both real adapters
satisfy it structurally with no inheritance change.

Migrated importers from `StatsDatabase` to `StatsRepository`:
  - tm1_integration.py
  - taskfile_ops.py
  - optimizer.py
  - contention_analyzer.py (replaces the AnyStatsDatabase Union shim;
    the alias remains for callers that imported it)
  - execution.py (docstring only)

db_admin.py is intentionally NOT migrated - it uses raw sqlite3
connections for backend-specific operations (VACUUM, raw queries) and
is by design SQLite-only.

Backwards compat preserved: ``from rushti.stats import StatsDatabase``
and every other historical import path still works via re-exports
in stats/__init__.py.

New test: tests/unit/test_stats_repository.py
  - Both real adapters satisfy the Protocol via isinstance()
  - A FakeStatsRepository also satisfies the Protocol
  - TaskOptimizer runs end-to-end against the fake (proof the seam
    is useful: optimizer tests can avoid real SQLite)

Suite size: 637 -> 642 tests, all passing.
